+++ src/main/java/com/project/cheerha/common/scheduler/core/TaskRepository.java
@@ -1,4 +1,4 @@
-package com.project.cheerha.common.repository;
+package com.project.cheerha.common.scheduler.core;
 
 import java.time.Instant;
